Improving the instructions
The first point is about git:
When i'm tried to install the repository in my project, i had a problem about cloned a repository into to another repository, so when i tried to do a push to project repository, the git had some errors. So i'm deleted the repository in the project, downloaded again and put to project folder, but not using the git clone and the problem is solved

The second point is about when the folder is the current name of this project (vtex-react-checkout):
In the first moment, i had not a folder 'checkout-ui-custom', so i put the repository folder into my project and try to run, but any changes don't resulted in workspace, so when i changed the folder name to 'checkout- ui-custom', its works.

So i changed some words and put more, to explain better this points.